Fixed HandledBoneCollected method
The issue was within SpawnBones()

I was removing the bone component from the object pool in the middle of the method, which caused syncing issues between what bones were activated vs which bones were in the activatedBones list.

Once I moved the code that removed the bone from the boolPool to execute last, HandledBoneCollected executed on the bone's OnCollected event.

When all bones have been collected, another set of bones are spawned.