Player locomotion configured
Player is configured with a free look virtual camera with a custom locomotion script for a third person controller.

The player can move with the "WSAD" keys, look around with the mouse, run with "SHIFT", and jump with "SPACE"

The player's character always faces the direction of the camera (ensuring the player is always moving where the user is looking), and by utilizing 'Quaternion.Slerp' the rotation of the player is smooth.

The custom Locomotion script also applies gravity to the character continuously unless grounded ensuring the jump motion is smooth.