Updated Field of View
Updated Field of View Visualization to be included within the Game View. Once mechanics are configured, the plan is to create a GUI system to turn on/off debug tools, as well as get on screen alerts/options to change AI States.

This project is starting to shift and I think I want to make a somewhat polished AI project. Allowing for GUI systems to showcase custom made debug tools and other aspects of AI that can be implemented into numerous other projects.